+++ R/utils.R
@@ -131,3 +131,24 @@ filter_function_args <- function(fun, args) {
   }
 }
 
+
+#' @title Assign attributes to an object from a named list
+#'
+#' @description
+#' Every list item will be made an attribute of the object.
+#' @param object An object to assign attributes to
+#' @param attribute_list A named list of attributes to assign to the object.
+#'
+#' @return The object with new attributes according to the contents of
+#' `attribute_list`
+#' @keywords internal
+assign_attributes <- function(object, attribute_list) {
+  if (is.null(object)) {
+    return(NULL)
+  }
+  for (i in seq_along(attribute_list)) {
+    setattr(object, names(attribute_list)[i], attribute_list[[i]])
+  }
+  return(object)
+}
+