+++ tests/testthat/test-summarise_scores.R
@@ -17,13 +17,15 @@ test_that("summarise_scores() works without any arguments", {
 
 test_that("summarise_scores() handles wrong by argument well", {
 
-  expect_error(summarise_scores(scores, by = "not_present"),
-    "The following items in `by` are notvalid column names of the data: 'not_present'. Check and run `summarise_scores()` again", # nolint
+  expect_error(
+    summarise_scores(scores, by = "not_present"),
+    "Column 'not_present' not found in data.", # nolint
     fixed = TRUE
   )
 
-  expect_error(summarise_scores(scores, by = "sample_id"),
-    "The following items in `by` are notvalid column names of the data: 'sample_id'. Check and run `summarise_scores()` again", # nolint
+  expect_error(
+    summarise_scores(scores, by = "sample_id"),
+    "Column 'sample_id' not found in data.",
     fixed = TRUE
   )
 })
@@ -34,58 +36,54 @@ test_that("summarise_scores() works with point forecasts in a quantile format", 
 
   scores <- suppressMessages(score(ex))
 
-  summarise_scores(scores, by = "model", na.rm = TRUE)
+  scores_summarised <-summarise_scores(scores, by = "model", na.rm = TRUE)
   expect_warning(
     expect_warning(
-      summarise_scores(
-        scores, by = "model", relative_skill = TRUE, na.rm = TRUE)
+      add_pairwise_comparison(scores_summarised)
     )
   )
 
   scores_point <- suppressMessages(score(example_point))
+  summarised_scores <- summarise_scores(scores_point, by = "model")
 
-  # expect_warning(
-  #   expect_warning(
-  #     summarise_scores(
-  #       scores_point, by = "model", relative_skill = TRUE, na.rm = TRUE)
-  #   )
-  # )
+  expect_no_condition(
+    pw_point <- add_pairwise_comparison(
+      summarised_scores,
+      relative_skill_metric = "se_point"
+    )
+  )
+
+  pw_manual <- pairwise_comparison(
+    scores_point, by = "model", metric = "se_point"
+  )
+
+  expect_equal(
+    pw_point$relative_skill,
+    unique(pw_manual$relative_skill)
+  )
 })
 
 test_that("summarise_scores() can compute relative measures", {
   ex <- data.table::copy(example_quantile)
   scores <- suppressMessages(score(ex))
 
-  expect_equal(
-    summarise_scores(
-      scores, by = "model", relative_skill = TRUE
-    )[, relative_skill],
-    c(1.6, 0.81, 0.75, 1.03), tolerance = 0.01
+  scores_with <- add_pairwise_comparison(
+    summarise_scores(scores, by = "model")
   )
 
   expect_equal(
-    summarise_scores(
-      scores, by = "model", relative_skill = TRUE,
-      relative_skill_metric = "ae_median"
-    )[, relative_skill],
+    scores_with[, relative_skill],
+    c(1.6, 0.81, 0.75, 1.03), tolerance = 0.01
+  )
+
+  scores_with <- add_pairwise_comparison(
+    summarise_scores(scores, by = "model"),
+    relative_skill_metric = "ae_median"
+  )
+
+  expect_equal(
+    scores_with[, relative_skill],
     c(1.6, 0.78, 0.77, 1.04), tolerance = 0.01
-  )
-})
-
-test_that("summarise_scores() metric is deprecated", {
-  ex <- data.table::copy(example_quantile)
-  scores <- suppressMessages(score(ex))
-
-  expect_equal(
-    suppressWarnings(summarise_scores(
-    scores, by = "model", metric = "auto", relative_skill = TRUE
-  ))[, relative_skill],
-    c(1.6, 0.81, 0.75, 1.03), tolerance = 0.01
-  )
-  expect_snapshot(
-    x <- summarise_scores(
-      scores, by = "model", metric = "auto", relative_skill = TRUE
-    )
   )
 })